fix(web): extract computeSourceStatus out of route file
Next.js disallows arbitrary named exports from route files — only HTTP
verb handlers and config fields are valid. Move computeSourceStatus to
lib/security/source-health-utils.ts; update route + test imports.
